ひっくり返す石を取得する処理をbase_boardに移動
